Tests/LibWeb/Text/expected/navigate-attachment-disposition-variants.txt

filename downloaded: true
uppercase downloaded: true
